config: move config to folder
