+++ index.ts
@@ -2,7 +2,7 @@
 import Sentry from "@sentry/node";
 import("./src/http");
 import Phoenix from "./src/Phoenix";
-import config from "./config.json" assert { type: "json" };
+import config from "./config/config.json" assert { type: "json" };
 
 Sentry.init({
     dsn: config.SENTRY_DSN,
